+++ packages/amazon-bedrock/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/amazon-bedrock
 
+## 4.0.55
+
+### Patch Changes
+
+- Updated dependencies [c60b393]
+  - @ai-sdk/anthropic@3.0.41
+
 ## 4.0.54
 
 ### Patch Changes