+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/google-vertex
 
+## 4.0.51
+
+### Patch Changes
+
+- Updated dependencies [3b3e32f]
+  - @ai-sdk/google@3.0.24
+
+## 4.0.50
+
+### Patch Changes
+
+- Updated dependencies [c60b393]
+  - @ai-sdk/anthropic@3.0.41
+
 ## 4.0.49
 
 ### Patch Changes